Refactor into PolymorphicTypesLookup
