+++ lib/jsonapi-resources.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'jsonapi/resources/railtie'
-require 'jsonapi/utils'
+require 'jsonapi/utils/polymorphic_types_lookup'
 require 'jsonapi/naive_cache'
 require 'jsonapi/compiled_json'
 require 'jsonapi/relation_retrieval'
